refactored callback system and init ddp
