+++ pytorch_lightning/accelerators/ddp_spawn_backend.py
@@ -103,7 +103,7 @@ class DDPSpawnBackend(Accelerator):
         # try to init for 20 times at max in case ports are taken
         # where to store ip_table
         model.trainer = self.trainer
-        model.init_ddp_connection(
+        self.init_ddp_connection(
             self.trainer.global_rank,
             self.trainer.world_size,
             self.trainer.is_slurm_managing_tasks
